rb: Refactor read_bytes into helper and add top-level API
DbgScript.read_bytes(addr, count) works like TypedObject.read_bytes(count)
except does not need any typed object. Caller passes address explicitly.

+++ src/support/util.cpp
@@ -293,7 +293,7 @@ UtilReadBytes(
 	_In_ DbgScriptHostContext* hostCtxt,
 	_In_ UINT64 addr,
 	_Out_writes_(cbCount) char* buf,
-	_In_ int cbCount,
+	_In_ ULONG cbCount,
 	_Out_ ULONG* cbActualLen)
 {
 	return hostCtxt->DebugDataSpaces->ReadVirtual(
